Fix/move catkin simple url to common defs

+++ modules/pull_auto_dependencies.sh
@@ -5,10 +5,12 @@ source $CI_MODULES/common_definitions.sh
 DEP_FILE_NAME=dependencies.rosinstall
 DEP_WORKSPACE_FILE=.rosinstall
 
+normalIFS="$IFS"
 IFS=$'\n' DEPS_FILES=($(find "${WORKSPACE}/src/" -maxdepth 4 -name $DEP_FILE_NAME | grep -Fv "^${WORKSPACE}/$DEPS/"))
+IFS="$normalIFS"
 
 if [[ ${#DEPS_FILES[@]} -eq 0 ]]; then
-  fatal "DEPENDENCIES = $DEPENDENCIES, but could not find any $DEP_FILE_NAME within ${WORKSPACE}/src/"
+  fatal "Could not find any $DEP_FILE_NAME within ${WORKSPACE}/src/"
 fi
 
 CHECKOUT_ASLAM_INSTALL=false
@@ -19,6 +21,7 @@ while true; do
   for dep_file in "${DEPS_FILES[@]}"; do
     echo "Processing dependency file $dep_file"
     IFS=$'\n' np=($($CI_MODULES/rosinstall-diff.py $DEP_WORKSPACE_FILE $dep_file))
+    IFS="$normalIFS"
     if [ -n "$np" ] ; then echo "Found new packages: ${np[@]}."; fi
     new_packages+=("${np[@]}")
     eval "$WSTOOL_MERGE_REPLACE $dep_file" # TODO Why is that necessary?
@@ -26,13 +29,18 @@ while true; do
   if [[ ${#new_packages[@]} -eq 0 ]] ; then 
     break
   fi
-  echo "Updating new packges "${new_packages[@]}""
+  echo "Updating new packges '${new_packages[@]}'"
   $WSTOOL_UPDATE_REPLACE "${new_packages[@]}"
-  IFS=$'\n' DEPS_FILES=($(find "${new_packages[@]}" -maxdepth 3 -name $DEP_FILE_NAME))
+  DEPS_FILES=($(find "${new_packages[@]}" -maxdepth 3 -name $DEP_FILE_NAME))
 done
 
 unset new_packages
 
 IFS=$'\n' all_superfluous_local_names=($($CI_MODULES/rosinstall-diff.py $DEP_WORKSPACE_FILE ${WORKSPACE}/$DEPS/))
-echo "Deleting superfluous ${all_superfluous_local_names[@]}"
-rm -irf "${all_superfluous_local_names[@]}"
+IFS="$normalIFS"
+
+if [ -n "$all_superfluous_local_names" ]; then
+  echo "Deleting superfluous ${all_superfluous_local_names[@]}"
+  rm -irf "${all_superfluous_local_names[@]}"
+fi
+
